@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class MinesweeperApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(MinesweeperApplication.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(MinesweeperApplication.class, args);
+    }
 }
