@@ -4,13 +4,13 @@ import com.deviget.minesweeper.model.dto.Cell;
 
 public class CellNotFoundException extends RuntimeException {
 
-	private static final long serialVersionUID = 4396573623121165297L;
+    private static final long serialVersionUID = 4396573623121165297L;
 
-	public CellNotFoundException(Cell cell) {
-		super("The cell " + cell + " was not found");
-	}
+    public CellNotFoundException(final Cell cell) {
+        super("The cell " + cell + " was not found");
+    }
 
-	public CellNotFoundException(Cell cell, Integer boardId) {
-		super("The cell " + cell + " was not found for the board " + boardId);
-	}
+    public CellNotFoundException(final Cell cell, final Integer boardId) {
+        super("The cell " + cell + " was not found for the board " + boardId);
+    }
 }
