@@ -3,9 +3,9 @@ package com.deviget.minesweeper.exception;
 public class ExistingUserException extends RuntimeException {
 
 
-	private static final long serialVersionUID = 5089524167351336546L;
+    private static final long serialVersionUID = 5089524167351336546L;
 
-	public ExistingUserException(String username) {
-		super("The user " + username + " already exists");
-	}
+    public ExistingUserException(final String username) {
+        super("The user " + username + " already exists");
+    }
 }
