@@ -2,9 +2,9 @@ package com.deviget.minesweeper.exception;
 
 public class GameOverException extends RuntimeException {
 
-	private static final long serialVersionUID = 4977580039388710504L;
+    private static final long serialVersionUID = 4977580039388710504L;
 
-	public GameOverException() {
-		super("The cell had a mine ! Game Over !");
-	}
+    public GameOverException() {
+        super("The cell had a mine ! Game Over !");
+    }
 }
