@@ -19,15 +19,15 @@ import java.util.List;
 @Table(name = "user")
 public class User extends BaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
 
-	@NotNull
-	@Column(unique = true)
-	private String username;
+    @NotNull
+    @Column(unique = true)
+    private String username;
 
-	@Builder.Default
-	@OneToMany(cascade = CascadeType.ALL, mappedBy = "user")
-	private List<Game> gameList = new ArrayList<>();
+    @Builder.Default
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "user")
+    private List<Game> gameList = new ArrayList<>();
 }
