@@ -4,6 +4,6 @@ package com.deviget.minesweeper.model.enums;
  * Enum for cell flag type representation.-
  */
 public enum FlagTypeEnum {
-	RED_FLAG,
-	QUESTION_MARK
+    RED_FLAG,
+    QUESTION_MARK
 }
