@@ -4,8 +4,8 @@ package com.deviget.minesweeper.model.enums;
  * Represents a Game Status.-
  */
 public enum GameStatusEnum {
-	IN_ṔROGRESS,
-	PAUSED,
-	COMPLETED,
-	FAILED
+    IN_PROGRESS,
+    PAUSED,
+    COMPLETED,
+    FAILED
 }
